Validate parameters
